Move PointModule to filters, fix copyright
